fix:change type of elements

+++ oxide_core/vector.rs
@@ -1,7 +1,7 @@
 use crate::matrix;
 
 pub trait Length {
-    fn get_length(&self) -> i32;
+    fn get_length(&self) -> f32;
 }
 
 pub trait Rotate {
@@ -10,7 +10,7 @@ pub trait Rotate {
 }
 
 pub trait InnerProduct {
-    fn inner_product(&self, _rhs: Self) -> i32;
+    fn inner_product(&self, _rhs: Self) -> f32;
 }
 
 pub trait OuterProduct {
