+++ oxide_core/vector3.rs
@@ -1,5 +1,5 @@
 pub struct Vector3 {
-    pub x: i32,
-    pub y: i32,
-    pub z: i32,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
